add login

+++ app/admin/action.ts
@@ -2,8 +2,11 @@
 
 import prisma from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
+import { auth } from "@/auth"
 
 export async function hapusPesanan(id: number) {
+  const session = await auth()
+  if (!session) throw new Error("Akses ditolak! Kamu bukan admin.")
   try {
     await prisma.order.delete({
       where: { id: id },
